refactor(common): table component code optimization
